renamed file to namibia.py
